Fixed circular dependency with constant by moving constants to separate file.
