+++ src/utils/resume/constants.ts
@@ -1,0 +1,10 @@
+export const SPACING = 10;
+export const VERTICAL_EDGE_SPACING = 100;
+export const HORIZONTAL_EDGE_SPACING = 60;
+export const RIGHT_COLUMN_START = 295;
+
+export const PAGE_WIDTH = 595.28;
+export const PAGE_HEIGHT = 841.89;
+export const COLUMN_WIDTH = (PAGE_WIDTH - HORIZONTAL_EDGE_SPACING * 3) / 2;
+
+export const MAX_TOP_SKILLS = 10;
