+++ src/utils/resume/resume.ts
@@ -2,8 +2,6 @@ import { PDFDocument } from 'pdf-lib';
 import IntroductionPage from './IntroductionPage';
 import SkillsPage from './SkillsPage';
 
-export const MAX_TOP_SKILLS = 10;
-
 export async function generateResume(): Promise<Uint8Array> {
     const pdfDoc = await PDFDocument.create();
